LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-cross-origin-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ')'

PASS window.onerror - compile error in <script src=//www1...>
PASS window.onerror - compile error in <script src=//www1...> (column)
